feat: move metadata into layout
